+++ dftt_timecode/__init__.py
@@ -1,5 +1,6 @@
 name = 'dftt_timecode'
 __author__ = 'You Ziyuan'
-__version__ = '0.0.12'
+__version__ = '0.0.13'
 
-from dftt_timecode.core import DfttTimecode
+from dftt_timecode.core.dftt_timecode import DfttTimecode
+from dftt_timecode.core.dftt_timerange import DfttTimeRange